[Ent Server 10.1.0] [EN-87912] Image Crop - Change DPI into double. (DB model change, WSDL change and Solr schema change.)
Former-commit-id: ab9b8d0da352bdf823e4044ab07079bc4b440fcb

+++ Enterprise/Enterprise/server/bizclasses/BizProperty.class.php
@@ -636,10 +636,10 @@ class BizProperty
 		self::$JoinProps['Height']  	 = null;
 		self::$JFldProps['Height']  	 = null;
 
-		self::$InfoProps['Dpi']          = new PropertyInfo( 'Dpi',         BizResources::localize( 'OBJ_DPI' ),        null,'int' );
+		self::$InfoProps['Dpi']          = new PropertyInfo( 'Dpi',         BizResources::localize( 'OBJ_DPI' ),        null,'double' );
 		self::$MetaProps['Dpi']          = 'ContentMetaData->Dpi';
 		self::$ObjFProps['Dpi']          = 'dpi';
-		self::$SqlTProps['Dpi']          = 'int';
+		self::$SqlTProps['Dpi']          = 'double';
 		self::$JoinProps['Dpi']  	 	 = null;
 		self::$JFldProps['Dpi']  	 	 = null;
