+++ Enterprise/Enterprise/server/dbclasses/DBVersion.class.php
@@ -23,7 +23,7 @@ class DBVersion extends DBBase
 			$versionFields = array(
 				'description' => 'blob', 'descriptionauthor' => 'string',  'keywords' => 'blob', 
 				'slugline' => 'string',  'format' => 'string', 
-				'_columns' => 'int', 'width' => 'double', 'depth' => 'double', 'dpi' => 'int', 
+				'_columns' => 'int', 'width' => 'double', 'depth' => 'double', 'dpi' => 'double',
 				'lengthwords' => 'int', 'lengthchars' => 'int', 'lengthparas' => 'int', 'lengthlines' => 'int',
 				'colorspace' => 'string', 'plaincontent' => 'blob', 'filesize' => 'int',
 				'comment' => 'string', 'types' => 'blob', 'modifier' => 'string', 'state' => 'int'
